fix(broker): issue with fmt and timestamp fixed
