+++ broker/unified_sql/inc/com/centreon/broker/unified_sql/stream.hh
@@ -54,22 +54,8 @@ struct uint64_not_null {
 
 }  // namespace unified_sql
 }  // namespace com::centreon::broker
+
 namespace fmt {
-template <>
-struct formatter<com::centreon::broker::timestamp> {
-  constexpr auto parse(format_parse_context& ctx)
-      -> format_parse_context::iterator {
-    return ctx.begin();
-  }
-
-  auto format(const com::centreon::broker::timestamp& t,
-              format_context& ctx) const -> format_context::iterator {
-    // ctx.out() is an output iterator to write to.
-    return t.is_null() ? fmt::format_to(ctx.out(), "NULL")
-                       : fmt::format_to(ctx.out(), "{}", t.get_time_t());
-  }
-};
-
 template <>
 struct formatter<com::centreon::broker::unified_sql::int64_not_minus_one> {
   constexpr auto parse(format_parse_context& ctx)
